Adding unit tests for email components

+++ services/intersection-api/api/src/main/java/us/dot/its/jpo/ode/api/controllers/users/UserController.java
@@ -66,7 +66,9 @@ public class UserController {
                 .map(EmailSubscription::getCategory)
                 .toList();
 
-        postgresService.removeEmailSubscriptionsByUser(userEmail, removedSubscriptions);
+        if (!removedSubscriptions.isEmpty()) {
+            postgresService.removeEmailSubscriptionsByUser(userEmail, removedSubscriptions);
+        }
         addedSubscriptions.forEach(subType -> {
             postgresService.addEmailSubscriptionByUser(userEmail, subType);
         });
